Stop building manylinux wheels
Wheels built in a CentOS-based manylinux docker container
won't work on Debian/Ubuntu Linux. This happens because the
wheels contain the libsasl2.so library which is not portable
across Linux distributions.

As far as I know there's no way to build CentOS-only wheels
so for now the safest thing to do is to stop distributing
wheels and keep the source distribution only.